requirements.txt
# decatrainer - DECA Flashcard Bank Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

